refactor: externalize handleTransaction
