+++ extensions/test-matchers/src/matchers/balance/toChangeBalances.spec.ts
@@ -423,7 +423,7 @@ describe('toChangeBalances', () => {
 				expect(error.message).toBe(
 					'Expected transaction to change balances by the specified amounts, but none of them passed',
 				)
-				console.log(error)
+
 				expect(error.actual).toMatchObject([
 					{
 						account: sender,
